Rename filterable-grid to gallery
